get main out of readsb.c (into readsb-main.c)
