+++ Makefile
@@ -71,7 +71,7 @@ readsb.pb-c.o: readsb.proto
 	protoc-c --c_out=. $<
 	$(CC) $(CPPFLAGS) $(CFLAGS) -c readsb.pb-c.c -o $@
 
-readsb: readsb.pb-c.o geomag.o readsb.o anet.o interactive.o mode_ac.o mode_s.o comm_b.o net_io.o crc.o demod_2400.o stats.o cpr.o icao_filter.o track.o util.o convert.o fifo.o sdr_ifile.o sdr_beast.o sdr.o ais_charset.o $(SDR_OBJ) $(COMPAT)
+readsb: readsb.pb-c.o geomag.o readsb.o readsb-main.o anet.o interactive.o mode_ac.o mode_s.o comm_b.o net_io.o crc.o demod_2400.o stats.o cpr.o icao_filter.o track.o util.o convert.o fifo.o sdr_ifile.o sdr_beast.o sdr.o ais_charset.o $(SDR_OBJ) $(COMPAT)
 	$(CC) -g -o $@ $^ $(LDFLAGS) $(LIBS) $(LIBS_SDR) 
 
 viewadsb: readsb.pb-c.o geomag.o viewadsb.o anet.o interactive.o mode_ac.o mode_s.o comm_b.o net_io.o crc.o stats.o cpr.o icao_filter.o track.o util.o ais_charset.o $(COMPAT)
